dsa: widen related k-cache kernel token ids
Use int64 token offsets in the matching DSA dequant and DSV4 quant kernels so large cache rows cannot overflow int32 pointer arithmetic.

+++ python/sglang/kernels/ops/attention/dsa/dequant_k_cache.py
@@ -132,7 +132,7 @@ def _dequantize_k_cache_fast_kernel(
     DIM_NOPE: tl.constexpr,
     DIM_ROPE: tl.constexpr,
 ):
-    token_id = tl.program_id(0)
+    token_id = tl.program_id(0).to(tl.int64)
     raw_block_id = tl.program_id(1)
 
     if raw_block_id < NUM_NOPE_BLOCKS:
